apply batch orchestrator patches (51 functions improved)
8 functions to 100%: AreUnlockableOutfitListConditionsMet,
ClearCrowds, GetSongShortName, IsLoaded<ObjDirPtr>,
vorbis_lsp_to_curve, _M_create_node, operator>>(MatChar),
DebugModal. Plus 43 partial improvements across flow, hamobj,
rndobj, math, net, stlport, and oggvorbis units.

+++ src/system/math/SHA1.cpp
@@ -34,7 +34,7 @@ static Licenses sLicense("system/src/math/SHA1.h", Licenses::kRequirementNotific
     w = rol(w, 30);
 
 void CSHA1::Transform(unsigned int *pState, const unsigned char *pBuffer) {
-    unsigned int a, b, c, d, e;
+    unsigned long a, b, c, d, e;
 
     a = pState[0];
     b = pState[1];
